requirements.txt
Pillow>=10.0,<11
tomli>=2.0.1 ; python_version < "3.11"
